refactor tests and add worker id to single hop path
